Efficiency improvement when deserializing (cRosMessageDeserialize()) a variable-length array of custom messages
When a received array is larger than the array size of the message that will store it, new elements of the array are created, but now the message definition files do not need to be loaded again when these new elements have to be created, that is, when the packet is deserialized. Instead the message definition (msgDef field) is used to create these elements.